requirements.txt
#
pytorch-lightning<2
omegaconf==2.2.3
#nerfacc==0.3.3
#
joblib==1.3.2
#av==10.0.0
numpy==1.23.0
chumpy==0.70
PyMCubes==0.1.2
pandas==2.0.2
dotmap==1.3.30
pyrender==0.1.45
trimesh==3.21.6
opencv-python==4.7.0.72
imageio==2.13.3
imageio-ffmpeg==0.4.8
scikit-image==0.18.1
scikit-learn==1.3.0
scipy==1.10.1
cython==3.0.6
av==11.0.0
#
smplx==0.1.28
tensorboard==2.13.0
lpips
#tinycudann==1.6
#Torch
#torch==1.13.1+cu116
torch-efficient-distloss==0.1.3
#torchmetrics==0.11.4
#torchsparse==1.4.0
#torchvision==0.14.1+cu116